Moving Group auth into Admin Console core code.
