+++ src/Api/Api.csproj
@@ -37,5 +37,5 @@
     <PackageReference Include="Azure.Messaging.EventGrid" Version="4.25.0" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.8.1" />
   </ItemGroup>
-
+  
 </Project>
